Add packwerk to dependencies
In the Gemfile, packwerk was specified, but this does not show that it depends on the packwerk gem when installed as a gem.

Therefore, it is necessary to describe the dependencies of packwerk gem in gemspec as well.
Note, however, that this gem depends on the implementation of the Pull Request, which is still under development,
so it will not work with the latest version of packwerk.